lps22hb: modify _from_lsb_to_hpa

+++ lps22hb_STdC/driver/lps22hb_reg.c
@@ -6,7 +6,7 @@
  ******************************************************************************
  * @attention
  *
- * <h2><center>&copy; COPYRIGHT(c) 2018 STMicroelectronics</center></h2>
+ * <h2><center>&copy; COPYRIGHT(c) 2019 STMicroelectronics</center></h2>
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -100,7 +100,7 @@ int32_t lps22hb_write_reg(lps22hb_ctx_t* ctx, uint8_t reg, uint8_t* data,
   *
   */
 
-float_t lps22hb_from_lsb_to_hpa(int16_t lsb)
+float_t lps22hb_from_lsb_to_hpa(uint32_t lsb)
 {
   return ( (float_t)lsb / 4096.0f );
 }
